schema: reviewers can be empty now + parsing: handle bold/italics formatting (#310)
* Change reviewers field to be optional

* Modify get_contributor_data return type to include None

Update get_contributor_data method to return None when no models are found.

* Add check for empty GitHub username and name

return None instead of trying to instantiate a ReviewUser

* fix models checking logic

* debug: hatch run

* go back to simpler handling of empty models

Refactor user model parsing to handle empty lists.

* Fix error handling for new contributors in update_review_teams

early return when we're missing a ReviewUser / fail to validate the PersonModel (skip behavior to align with the existing log statement)

* Add '*' to unwanted characters in clean_name function

there's a contributor that's being processed with * in their name. stripping this character should allow them to be processed as a PersonModel (in theory)

* Remove markdown formatting from issue value

Strip markdown formatting from returned value.

* Update CHANGELOG for version 1.7.6

Added details for version 1.7.6 including bug fix and feature.